backend/requirements.txt
fastapi>=0.115
uvicorn>=0.30
openai>=1.30
scikit-learn>=1.4
python-dotenv>=1.0
pydantic>=2.7
